# map: avoid allocations during PerCPUMap batch lookups