@lmb 2025-07-18
Maybe having it in the commit message is fine then!